requirements.txt
numpy~=1.19.2
matplotlib~=3.3.3
pandas~=1.1.5
torch>=1.0.0
torchvision>=0.2.2
tqdm~=4.56.0
scipy~=1.5.4
faiss-gpu
mkl
mkl-service
accimage
pillow~=8.1.0